Added `debug` config option to output zone data
I don't have access to an Android device and thus can't use the FloleVac option for determining the zones in my house. This change allows zones to be drawn using the mode: `Zoned cleanup` and the resulting coordinates to be shown on screen, so they can be copied and pasted into the YML config for the card. Requires setting `debug: true` in config first.